optimized AddressLine with recompose for pure render.
