+++ src/containers/checkout/components/index.js
@@ -7,10 +7,8 @@ export ProductReview from './ProductReview/';
 export GrandTotal from './GrandTotal/';
 export SubmitOrder from './SubmitOrder/';
 export NetworkStatus from './NetworkStatus/';
-export AddressLine from './ShippingAddress/addressLine';
 export Country from './ShippingAddress/country';
 export Prefecture from './ShippingAddress/prefecture';
-export PostalCode from './ShippingAddress/postalCode';
 export City from './ShippingAddress/city';
 export PhoneNumber from './ShippingAddress/phoneNumber';
 export CvnModal from './CreditCardInfo/components/CvnModal';
@@ -19,4 +17,6 @@ export {
   GivenName,
   FamilyName,
   Email,
+  PostalCode,
+  AddressLine,
 } from './ShippingAddress/components/';
