+++ src/containers/checkout/components/productReview/index.js
@@ -42,8 +42,7 @@ function ProductReview({
 }
 const ProductReviewWithLifecycle = lifecycle({
   shouldComponentUpdate(nextProps) {
-    const npCopy = _.cloneDeep(nextProps);
-    const tpCopy = _.cloneDeep(this.props);
+
 
     if (!_.isEqual(npCopy, tpCopy)) return true;
     return false;
